requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
openai>=1.51.0
supabase==2.7.4
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.5.2
httpx==0.27.2
python-telegram-bot==21.10
requests==2.32.3
